Create a Hierarchical Dirichlet Mixture of Beta Distributions

Create a Hierarchical Dirichlet Mixture of Beta Distributions

DirichletProcessHierarchicalBeta(dataList, maxY, priorParameters = c(2,
  8), hyperPriorParameters = c(1, 0.125), gammaPriors = c(2, 4),
  alphaPriors = c(2, 4), mhStepSize = c(0.1, 0.1), numSticks = 50,
  mhDraws = 250)

Arguments

dataList
List of data for each separate Dirichlet mixture object

maxY
Maximum value for the Beta distribution.

priorParameters
Prior Parameters for the top level base distribution.

hyperPriorParameters
Hyper prior parameters for the top level base distribution.

gammaPriors
Prior parameters for the top level concentration parameter.

alphaPriors
Prior parameters for the individual parameters.

mhStepSize
Metropolis Hastings jump size.

numSticks
Truncation level for the Stick Breaking formulation.

mhDraws
Number of Metropolis-Hastings samples to perform for each cluster update.

Value

dpobjlist A Hierarchical Dirichlet Process object that can be fitted, plotted etc.