Create a Dirichlet Mixture of the Weibull distribution

The likelihood is parameterised as \(\mathrm{Weibull} (y | a, b) = \frac{a}{b} y ^{a-1} \exp \left( - \frac{x^a}{b} \right)\). The base measure is a Uniform Inverse Gamma Distribution. \(G_0 (a, b | \phi, \alpha _0 , \beta _0) = U(a | 0, \phi ) \mathrm{Inv-Gamma} ( b | \alpha _0, \beta _0)\) \(\phi \sim \mathrm{Pareto}(x_m , k)\) \(\beta \sim \mathrm{Gamma} (\alpha _0 , \beta _0)\) This is a semi-conjugate distribution. The cluster parameter a is updated using the Metropolis Hastings algorithm an analytical posterior exists for b.

DirichletProcessWeibull(y, g0Priors, alphaPriors = c(2, 4),
  mhStepSize = c(1, 1), hyperPriorParameters = c(6, 2, 1, 0.5),
  verbose = FALSE, mhDraws = 250)

Arguments

y
Data.

g0Priors
Base Distribution Priors.

alphaPriors
Prior for the concentration parameter.

mhStepSize
Step size for the new parameter in the Metropolis Hastings algorithm.

hyperPriorParameters
Hyper prior parameters.

verbose
Set the level of screen output.

mhDraws
Number of Metropolis-Hastings samples to perform for each cluster update.

Value

Dirichlet process object

References

Kottas, A. (2006). Nonparametric Bayesian survival analysis using mixtures of Weibull distributions. Journal of Statistical Planning and Inference, 136(3), 578-596.