Dirichlet process mixture of the Beta distribution.

Create a Dirichlet process object using the mean and scale parameterisation of the Beta distribution bounded on \((0, maxY)\).

DirichletProcessBeta(y, maxY, g0Priors = c(2, 8), alphaPrior = c(2, 4),
  mhStep = c(1, 1), hyperPriorParameters = c(1, 0.125),
  verbose = TRUE, mhDraws = 250)

Arguments

y
Data for which to be modelled.

maxY
End point of the data

g0Priors
Prior parameters of the base measure \((\alpha _0, \beta _0)\).

alphaPrior
Prior parameters for the concentration parameter. See also UpdateAlpha.

mhStep
Step size for Metropolis Hastings sampling algorithm.

hyperPriorParameters
Hyper-prior parameters for the prior distributions of the base measure parameters \((a, b)\).

verbose
Logical, control the level of on screen output.

mhDraws
Number of Metropolis-Hastings samples to perform for each cluster update.

Value

Dirichlet process object

Details

\(G_0 (\mu , \nu | maxY, \alpha _0 , \beta _0) = U(\mu | 0, maxY) \mathrm{Inv-Gamma} (\nu | \alpha _0, \beta _0)\).

The parameter \(\beta _0\) also has a prior distribution \(\beta _0 \sim \mathrm{Gamma} (a, b)\) if the user selects Fit(...,updatePrior=TRUE).